mejoras 310122

+++ database/migrations/2022_01_26_172213_create_detalleordenservicio_table.php
@@ -24,7 +24,8 @@ class CreateDetalleordenservicioTable extends Migration
             $table->integer('cantidadKardex');
             $table->unsignedBigInteger('unidmedida_id')->required();
             $table->foreign('unidmedida_id')->references('id')->on('unidmedida');
-            $table->decimal('punit', 8, 4)->required();
+            /* $table->decimal('punit', 8, 4)->required(); */
+            $table->float('punit')->required();
             $table->enum('estado',[Ordenservicio::ATENDIDO,Ordenservicio::ANULADO])->default(Ordenservicio::ATENDIDO);
         });
     }
